images/transcribe/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohttp==3.9.3
    # via fsspec
aiosignal==1.3.1
    # via aiohttp
alembic==1.13.1
    # via optuna
antlr4-python3-runtime==4.9.3
    # via omegaconf
asteroid-filterbanks==0.4.0
    # via pyannote-audio
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via aiohttp
audioread==3.0.1
    # via librosa
boto3==1.34.54
    # via -r requirements.in
botocore==1.34.54
    # via
    #   boto3
    #   s3transfer
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via soundfile
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via typer
colorama==0.4.6
    # via typer
colorlog==6.8.2
    # via optuna
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
decorator==5.1.1
    # via librosa
docopt==0.6.2
    # via
    #   pyannote-metrics
    #   pyannote-pipeline
einops==0.7.0
    # via pyannote-audio
ffmpeg-python==0.2.0
    # via -r requirements.in
filelock==3.13.1
    # via
    #   huggingface-hub
    #   pyannote-pipeline
    #   torch
    #   transformers
fonttools==4.49.0
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2024.2.0
    # via
    #   huggingface-hub
    #   lightning
    #   pytorch-lightning
    #   torch
future==1.0.0
    # via ffmpeg-python
huggingface-hub==0.21.3
    # via
    #   pyannote-audio
    #   speechbrain
    #   tokenizers
    #   transformers
hyperpyyaml==1.2.2
    # via speechbrain
idna==3.6
    # via
    #   requests
    #   yarl
jinja2==3.1.3
    # via torch
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.3.2
    # via
    #   librosa
    #   scikit-learn
    #   speechbrain
julius==0.2.7
    # via torch-audiomentations
kiwisolver==1.4.5
    # via matplotlib
lazy-loader==0.3
    # via librosa
librosa==0.10.1
    # via torch-audiomentations
lightning==2.2.0.post0
    # via pyannote-audio
lightning-utilities==0.10.1
    # via
    #   lightning
    #   pytorch-lightning
    #   torchmetrics
llvmlite==0.42.0
    # via numba
mako==1.3.2
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
matplotlib==3.8.3
    # via pyannote-metrics
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
msgpack==1.0.8
    # via librosa
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
networkx==3.2.1
    # via torch
numba==0.59.0
    # via librosa
numpy==1.26.4
    # via
    #   -r requirements.in
    #   asteroid-filterbanks
    #   contourpy
    #   librosa
    #   lightning
    #   matplotlib
    #   numba
    #   optuna
    #   pandas
    #   pyannote-core
    #   pyannote-metrics
    #   pytorch-lightning
    #   pytorch-metric-learning
    #   scikit-learn
    #   scipy
    #   soxr
    #   speechbrain
    #   tensorboardx
    #   torchmetrics
    #   transformers
omegaconf==2.3.0
    # via pyannote-audio
optuna==3.5.0
    # via pyannote-pipeline
packaging==23.2
    # via
    #   huggingface-hub
    #   lightning
    #   lightning-utilities
    #   matplotlib
    #   optuna
    #   pooch
    #   pytorch-lightning
    #   speechbrain
    #   tensorboardx
    #   torch-pitch-shift
    #   torchmetrics
    #   transformers
pandas==2.2.1
    # via
    #   pyannote-database
    #   pyannote-metrics
pillow==10.2.0
    # via matplotlib
platformdirs==4.2.0
    # via pooch
pooch==1.8.1
    # via librosa
primepy==1.3
    # via torch-pitch-shift
protobuf==4.25.3
    # via tensorboardx
pyannote-audio==3.1.1
    # via -r requirements.in
pyannote-core==5.0.0
    # via
    #   pyannote-audio
    #   pyannote-database
    #   pyannote-metrics
    #   pyannote-pipeline
pyannote-database==5.0.1
    # via
    #   pyannote-audio
    #   pyannote-metrics
    #   pyannote-pipeline
pyannote-metrics==3.2.1
    # via pyannote-audio
pyannote-pipeline==3.0.1
    # via pyannote-audio
pycparser==2.21
    # via cffi
pydub==0.25.1
    # via -r requirements.in
pygments==2.17.2
    # via rich
pyodbc==5.1.0
    # via -r requirements.in
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   matplotlib
    #   pandas
pytorch-lightning==2.2.0.post0
    # via lightning
pytorch-metric-learning==2.4.1
    # via pyannote-audio
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   hyperpyyaml
    #   lightning
    #   omegaconf
    #   optuna
    #   pyannote-database
    #   pyannote-pipeline
    #   pytorch-lightning
    #   transformers
regex==2023.12.25
    # via transformers
requests==2.31.0
    # via
    #   huggingface-hub
    #   pooch
    #   transformers
rich==13.7.1
    # via
    #   pyannote-audio
    #   typer
ruamel-yaml==0.18.6
    # via hyperpyyaml
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
s3transfer==0.10.0
    # via boto3
safetensors==0.4.2
    # via transformers
scikit-learn==1.4.1.post1
    # via
    #   librosa
    #   pyannote-metrics
    #   pyannote-pipeline
    #   pytorch-metric-learning
scipy==1.12.0
    # via
    #   librosa
    #   pyannote-core
    #   pyannote-metrics
    #   scikit-learn
    #   speechbrain
semver==3.0.2
    # via pyannote-audio
sentencepiece==0.2.0
    # via speechbrain
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sortedcontainers==2.4.0
    # via pyannote-core
soundfile==0.12.1
    # via
    #   librosa
    #   pyannote-audio
soxr==0.3.7
    # via librosa
speechbrain==1.0.0
    # via pyannote-audio
sqlalchemy==2.0.27
    # via
    #   alembic
    #   optuna
sympy==1.12
    # via
    #   pyannote-metrics
    #   torch
tabulate==0.9.0
    # via pyannote-metrics
tensorboardx==2.6.2.2
    # via pyannote-audio
threadpoolctl==3.3.0
    # via scikit-learn
tokenizers==0.15.2
    # via transformers
torch==2.2.1
    # via
    #   -r requirements.in
    #   asteroid-filterbanks
    #   julius
    #   lightning
    #   pyannote-audio
    #   pytorch-lightning
    #   pytorch-metric-learning
    #   speechbrain
    #   torch-audiomentations
    #   torch-pitch-shift
    #   torchaudio
    #   torchmetrics
torch-audiomentations==0.11.1
    # via pyannote-audio
torch-pitch-shift==1.2.4
    # via torch-audiomentations
torchaudio==2.2.1
    # via
    #   pyannote-audio
    #   speechbrain
    #   torch-audiomentations
    #   torch-pitch-shift
torchmetrics==1.3.1
    # via
    #   lightning
    #   pyannote-audio
    #   pytorch-lightning
tqdm==4.66.2
    # via
    #   huggingface-hub
    #   lightning
    #   optuna
    #   pyannote-pipeline
    #   pytorch-lightning
    #   pytorch-metric-learning
    #   speechbrain
    #   transformers
transformers==4.38.2
    # via -r requirements.in
typer[all]==0.9.0
    # via
    #   pyannote-database
    #   typer
typing-extensions==4.10.0
    # via
    #   alembic
    #   asteroid-filterbanks
    #   huggingface-hub
    #   librosa
    #   lightning
    #   lightning-utilities
    #   pyannote-core
    #   pytorch-lightning
    #   sqlalchemy
    #   torch
    #   typer
tzdata==2024.1
    # via pandas
urllib3==2.0.7
    # via
    #   botocore
    #   requests
yarl==1.9.4
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
